Creating two builders for character and weapon

+++ HeroesAndCreatureUnitTesting/Consts/WeaponConsts.cs
@@ -1,5 +1,5 @@
 ﻿namespace UnitTests.Consts {
     public static class WeaponConsts {
-        public static readonly double DefaultWeaponPower = 10.0;
+        public static readonly float DefaultWeaponPower = 10.0f;
     }
 }
