+++ HeroesAndCreatures/Weapons/IWeapon.cs
@@ -1,6 +1,6 @@
 ﻿namespace HeroesAndCreatures.Weapons {
     public interface IWeapon {
         string Name { get; }
-        double Attack();
+        float Attack();
     }
 }
